Rename init to scaffold
